consolidate plugin code
