+++ grouper/plugin.py
@@ -4,6 +4,30 @@ plugin.py
 Base plugin for Grouper plugins. These are plugins that can be written to extend Grouper
 functionality.
 """
+from annex import Annex
+
+
+Plugins = []
+
+
+class PluginsAlreadyLoaded(Exception):
+    pass
+
+
+def load_plugins(plugin_dir, service_name):
+    """Load plugins from a directory"""
+    global Plugins
+    if Plugins:
+        raise PluginsAlreadyLoaded("Plugins already loaded; can't load twice!")
+    Plugins = Annex(BasePlugin, [plugin_dir], raise_exceptions=True)
+    for plugin in Plugins:
+        plugin.configure(service_name)
+
+
+def get_plugins():
+    """Get a list of loaded plugins."""
+    global Plugins
+    return list(Plugins)
 
 
 class BasePlugin(object):
